More progress changing UInt64 => IntPtr where needed
Former-commit-id: 7e3ce8597d4fa0b27fdc1cdc53891ecdfd9000b5

+++ Anathema/Source/Tools/Debugger/Memory View/IMemoryViewMVP.cs
@@ -14,7 +14,7 @@ namespace Anathema
     class MemoryViewEventArgs : EventArgs
     {
         public IEnumerable<NormalizedRegion> VirtualPages;
-        public UInt64 Address;
+        public IntPtr Address;
     }
 
     interface IMemoryViewView : IView
@@ -22,7 +22,7 @@ namespace Anathema
         // Methods invoked by the presenter (upstream)
         void ReadValues();
         void UpdateVirtualPages(List<String> VirtualPages);
-        void GoToAddress(UInt64 Address);
+        void GoToAddress(IntPtr Address);
     }
 
     abstract class IMemoryViewModel : RepeatedTask, IModel
